container_imgs/scrapper/requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
setuptools==74.1.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
boto3==1.35.15
unidecode==1.3.8
